Update interaction handlers to use inventory crate
The previous implementation, which defined all handlers in the feather-server-player crate,
would require feather-server-player to implement functionality related to chests
which belongs in the feather-server-block crate. Using inventory allows interaction
handlers to be defined in different crates.

Note that this has temporarily removed the crafting table handler, but it will
be added back once crafting is implemented.
