+++ server/block/src/chest.rs
@@ -6,14 +6,19 @@ use feather_core::{
         player::InventorySlot,
     },
     blocks::{BlockKind, ChestKind, FacingCardinal},
-    inventory::Area,
+    inventory::{Area, Window},
     items::{Item, ItemStack},
-    network::{packets::BlockAction, Packet},
+    network::{
+        packets::{BlockAction, OpenWindow, WindowItems},
+        Packet,
+    },
+    text::TextRoot,
 };
 use feather_server_entity::drops::drop_item;
 use feather_server_types::{
     BlockEntityLoaderRegistration, BlockSerializer, BlockUpdateCause, BlockUpdateEvent, BumpVec,
-    Game, Inventory, SpawnPacketCreator, WindowCloseEvent, WindowOpenEvent,
+    Game, InteractionHandler, Inventory, Network, SpawnPacketCreator, WindowCloseEvent,
+    WindowOpenEvent,
 };
 use fecs::{Entity, EntityBuilder, EntityRef, World};
 use num_traits::ToPrimitive;
@@ -291,6 +296,60 @@ fn connected_offset(facing: FacingCardinal) -> [i32; 2] {
         FacingCardinal::South => [1, 0],
         FacingCardinal::East => [0, -1],
         FacingCardinal::West => [0, 1],
+    }
+}
+
+/// Handler for player right clicking on chests.
+pub struct ChestInteraction;
+inventory::submit!(Box::new(ChestInteraction) as Box<dyn InteractionHandler>);
+
+impl InteractionHandler for ChestInteraction {
+    fn handle_interaction(
+        &self,
+        game: &mut Game,
+        world: &mut World,
+        pos: BlockPosition,
+        player: Entity,
+        window_id: u8,
+    ) {
+        // Open chest window and set the player's window
+        if let Some(chest_entity) = game.block_entities.get(&pos).copied() {
+            let packet = OpenWindow {
+                window_id,
+                window_type: String::from("minecraft:chest"),
+                window_title: TextRoot::from("Chest").into(),
+                number_of_slots: 27,
+                entity_id: None,
+            };
+            world.get::<Network>(player).send(packet);
+
+            {
+                let mut slots = Vec::new();
+                let chest = world.get::<Inventory>(chest_entity);
+                for i in 0..27 {
+                    slots.push(chest.item_at(Area::Chest, i).unwrap());
+                }
+
+                let packet = WindowItems { window_id, slots };
+                world.get::<Network>(player).send(packet);
+            }
+
+            world
+                .add(player, Window::chest(player, chest_entity))
+                .unwrap();
+
+            game.handle(
+                world,
+                WindowOpenEvent {
+                    player,
+                    opened: chest_entity,
+                },
+            );
+        }
+    }
+
+    fn block_kind(&self) -> BlockKind {
+        BlockKind::Chest
     }
 }
