+++ server/player/Cargo.toml
@@ -23,3 +23,4 @@ ahash = "0.3"
 parking_lot = "0.10"
 thiserror = "1.0"
 anyhow = "1.0"
+inventory = "0.1"
