+++ server/player/src/packet_handlers.rs
@@ -1,7 +1,6 @@
 //! Systems which handle packets.
 
 mod animation;
-mod block_interaction;
 mod chat;
 mod client_status;
 mod digging;
@@ -11,12 +10,12 @@ mod placement;
 mod use_item;
 mod window;
 
+pub use self::inventory::*;
 pub use animation::handle_animation;
 pub use chat::handle_chat;
 pub use client_status::handle_client_status;
 pub use digging::*;
 use fecs::{Entity, World};
-pub use inventory::*;
 pub use movement::handle_movement_packets;
 pub use placement::handle_player_block_placement;
 pub use use_item::handle_player_use_item;
